backend/requirements.txt
Flask>=2.0
Flask-SocketIO>=5.0
Flask-Cors>=3.0
python-dotenv>=0.19
eventlet 